backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.18
pillow==11.0.0
pydantic==2.10.3
python-dotenv==1.0.1
aiofiles==24.1.0
replicate==0.25.1
